pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-mcp-server>=0.1.1",
    "openai-agents[litellm]>=0.0.16",
    "python-dotenv>=1.1.0",
]
